+++ packages/adhere-util-communication-ajax/lib/ajax.d.ts
@@ -1,4 +1,4 @@
-import { IConfig, ISendArg } from './types';
+import type { IConfig, ISendArg } from './types';
 /**
  * Ajax
  * @class Ajax